Adjusted Speedster acceleration and added some comments, mostly for me
Increased Speedster acceleration so that it is a faster class. Idk, all
of these stats can be adjusted later when we play test.

I added a few comments, some at the top of the player classes bc I had
forgotten those and then one in the Game class which reminds me that I
had set up the classes correctly (YAY).