(Attempted to) Implement collisions
Added method in Game state to check for collisions in between objects
and, if so, to call .onCollision on the objects. The Game currently
calculates the resulting velocities of the two objects impacting however
this implementation may be incorrect. It is difficult to determine if it
is correct or not because collision detection is not exactly working.

Implementation of this uses PolyBounds and its .intersect method to
determine if two objects collided or not. Idk if this is what that
method even does to be honest but this can be replaced by some other
method that determines if the objects have collided.