Collision testing
The reason the collisions aren't working currently is because, for the
player object, it's sprite's Poly is being used to check for
intersections but the points used for that check aren't adjusted for the
player's current position on the screen.

Idk exactly all that I changed, it was last night and this morning and I
can't see the change log until I commit, so I may have to do some
cleaning up.