Fixed collision-checking
Reworked Poly class to use the object's position in the game when
checking for collisions. 

The player classes use the polyBounds() gotten from sprite.get(0) so the
collision bounds are based off of whatever part of the sprite that
corresponds to. For BasicClass ("podracer"), it is that inner part so
that inner part has to collide with the ball in order to trigger a
collision.

Now we gotta figure out what happens after a collision. For the ball,
how does it move.