Added goals and reset ball after collision with one
Added basic goals. Idk where exactly we want to put them, I decided to
put them just inside the FieldObjects but they can be moved to be
against the FieldObject's left wall.

Added reset method to ball to be able to reset after a shot. It resets
its position, velocity, and acceleration.