Added WASD player controls, made player's angle based off mouse, and +
Added WASD player controls in the playerMovement() method (BasicClass)
and added mouseAngle() and playerAngle() in order to make it so that if
the bool isPlayer is true, that the player looks at the mouse.

Made it so that drag and recoil works like actual drag and recoil now.

Idk what this .gitignore is but it's gone.