Adjusted methods to generalize Player class better
Made 'checkDrag' and 'shoot' methods that are called inside
'updatePhysics' and 'playerMovement' respectively. This is so that, if
'isPlayer' is false (maybe is a bot), it can just use the 'shoot' method
do shoot. Checking for drag is independent of whether or not it is a
player or not so it is moved into the more general 'updatePhysics'.

Also, I guess in implementing these I changed something and now the
mouse movement is more physics-based? As in, velocity isn't just changed
to be the angle of the mouse; it is based off of acceleration which is
generated from pressing space.

TL;DR moved stuff and now the player can drift bc physics I guess