Implemented right-click handling and added specials to BasicClass
In Input, added lastRightMousePress and lastRightMouseRelease among
other things that follow handling of regular mouse click.
Modified mousePressed() and mouseReleased() in order to do so.

Added a useSpecial() method in BasicClass that is called if the right
mouse is pressed. Currently, it just shoots a large missile but this can
be changed to do whatever later.