Added a game timer
Added HUD object that is added to the object list inside of the Game
gamestate. 

Added NumberCharacters enum that holds all of the digits (0-9) and colon
and decimal point for scoring and the game timer and whatnot.

Added GameTimer object that uses a Java Swing Timer to handle the game
timer.