Edited projectile's constructor for correct velocity and position
Missiles were not shooting in the correct direction (they were just
going directly to the right) so that the heading given to the Missile in
its constructor was then sent to the projectile to use.

Also, the missiles were constantly stuck at the origin so its position
was updated to use the position that the player was at when it was shot
(also given to Missile constructor).