Fixed weird player collision and graphics bug; Made ball round
In 'trigger' method, set 'poly' to equal 'polyBounds().clone()' so that
a cloned version is used rather than making reference to the actual
first poly and throwing things off by changing that offset. Basically,
the center poly of the player (the one used for collision detection) was
flying off to places and was acting weird and inverted so it was fixed.

The ball is now round. New sprite which is a circle. Because balls are
round. 

Also, I changed the podracer scale to be 50x greater (the scale used in
'.scaleNew' inside of the BasicClass' constructor) because I initially
thought that that would help with the weird collisions and graphics
stuff but it didn't and I don't really want to change it by hand. It
isn't broken so...