I'm gonna be honest, I don't really remember what I did
I tried to fix the missiles in that they aren't showing for some reason.
I didn't get anywhere with that.

I also made the ball smaller and made it grey so that it contrasts with
the field objects.

The actual goal objects where made really thin bc apparently they are
supposed to be essentially outside of the bounds like in real soccer
(the field objects are just markers for the goal box or something).

BMath was changed. Idk really what was added. I think it was a clamp for
like doubles?

BasicClass was altered in a WIP attempt to add AI.